backend/requirements.txt
langchain
langchain-anthropic
langchain-community
langchain-google-genai
langchain-google-vertexai
langchain-openai
langchain-tavily
langgraph
langgraph-cli[inmem]
pydantic
python-dotenv
requests
markdown
pillow
fastapi
uvicorn[standard]
vertexai
google-cloud-aiplatform
